# feat(duplication): make the task code for incremental loading from private logs configurable