--- a/src/replica/duplication/replica_duplicator.cpp
+++ b/src/replica/duplication/replica_duplicator.cpp
@@ -170,12 +170,14 @@ void replica_duplicator::start_dup_log()
     _load = std::make_unique<load_mutation>(this, _replica, _load_private.get());
 
     from(*_load).link(*_ship).link(*_load);
-    const dsn::task_code load_from_private_log_level =
-        dsn::task_code::try_get(FLAGS_load_from_private_log_level, LPC_REPLICATION_LONG_LOW);
-    LOG_ERROR_PREFIX("unexpected load_from_private_log_level ({}), put it to default value "
-                     "LPC_REPLICATION_LONG_LOW",
-                     load_from_private_log_level);
-    fork(*_load_private, load_from_private_log_level, 0).link(*_ship);
+    auto dup_load_plog_task =
+        dsn::task_code::try_get(FLAGS_dup_load_plog_task, TASK_CODE_INVALID);
+    if (dup_load_plog_task == TASK_CODE_INVALID) {
+        dup_load_plog_task = LPC_REPLICATION_LONG_LOW;
+        LOG_ERROR_PREFIX("invalid dup_load_plog_task ({}), set it to LPC_REPLICATION_LONG_LOW",
+                                              FLAGS_dup_load_plog_task);
+    }
+    fork(*_load_private, dup_load_plog_task, 0).link(*_ship);
 
     run_pipeline();
 }