--- a/src/replica/duplication/replica_duplicator.cpp
+++ b/src/replica/duplication/replica_duplicator.cpp
@@ -46,11 +46,12 @@ METRIC_DEFINE_counter(replica,
                       dsn::metric_unit::kMutations,
                       "The number of confirmed mutations for dup");
 
-DSN_DEFINE_string(replication,
-                  dup_load_plog_task,
-                  "LPC_REPLICATION_LONG_LOW",
-                  "The task code for incremental loading from private logs while duplicating. Tasks with "
-                  "TASK_PRIORITY_HIGH are not recommended.");
+DSN_DEFINE_string(
+    replication,
+    dup_load_plog_task,
+    "LPC_REPLICATION_LONG_LOW",
+    "The task code for incremental loading from private logs while duplicating. Tasks with "
+    "TASK_PRIORITY_HIGH are not recommended.");
 DSN_TAG_VARIABLE(dup_load_plog_task, FT_MUTABLE);
 
 namespace dsn::replication {
@@ -169,12 +170,11 @@ void replica_duplicator::start_dup_log()
     _load = std::make_unique<load_mutation>(this, _replica, _load_private.get());
 
     from(*_load).link(*_ship).link(*_load);
-    auto dup_load_plog_task =
-        dsn::task_code::try_get(FLAGS_dup_load_plog_task, TASK_CODE_INVALID);
+    auto dup_load_plog_task = dsn::task_code::try_get(FLAGS_dup_load_plog_task, TASK_CODE_INVALID);
     if (dup_load_plog_task == TASK_CODE_INVALID) {
         dup_load_plog_task = LPC_REPLICATION_LONG_LOW;
         LOG_ERROR_PREFIX("invalid dup_load_plog_task ({}), set it to LPC_REPLICATION_LONG_LOW",
-                                              FLAGS_dup_load_plog_task);
+                         FLAGS_dup_load_plog_task);
     }
     fork(*_load_private, dup_load_plog_task, 0).link(*_ship);