### What problem does this PR solve? <!--add issue link with summary if exists-->

https://github.com/apache/incubator-pegasus/issues/2183

### What is changed and how does it work?

We can make the task code configurable, allowing the thread priority incremental
loading from private logs of to be adjusted from **LOW** to **COMMON**, thereby
enabling support for low-latency real-time synchronization.


##### Performance Testing <!-- At least one of them must be included. -->
I do some test cases as following:
In these test cases, I first wrote 8k QPS of write traffic to the master cluster (this is the traffic that my test cluster will not generate dup log backlogs) to verify the effect of the priority modification. Then I wrote 20k QPS of write traffic to the master cluster (this is the traffic that my test cluster will generate a certain degree of dup log backlogs) to verify the effect of the priority modification.

| load/ship task priority | load_from_private_log task priority | qps  | duplicate_log_batch_bytes | plog Maximum backlog | Master cluster write delay p99 | master/slave dup delay     |
| -------------------- | -------------------------------- | ---- | ------------------------- | ----------- | ---------------- | ------------------------ |
| LOW                  | LOW                              | 8K   | 4096                      | 9k          | 1ms              | p95 127ms、p99 27473ms   |
| LOW                  | COMMON                           | 8K   | 4096                      | 200         | 1ms              | p95 101ms、p99 109ms     |
| HIGH                 | HIGH                             | 8K   | 4096                      | 150         | 1ms              | p95 107ms、p99 115ms     |
| LOW                  | LOW                              | 20K  | 4096                      | 61K         | 1.5ms            | p95 139ms、p99 20506ms   |
| LOW                  | COMMON                           | 20K  | 4096                      | 42K         | 1.5ms            | p95 126ms、p99 18127ms   |
| LOW                  | COMMON                           | 20K  | 0                         | Continue to increase over time    | 1.5ms            | 95 10618ms、p99 303519ms |


As you see , change the priority from LOW to COMMON of `load_from_private_log` will not t increase the online delay. And priority from LOW to HIGH is no benefit for further speeding up duplication.

So based on the above experimental conclusions, I think this issues' argument is valid.